+++ RELEASE_NOTES.md
@@ -17,6 +17,8 @@ Breaking changes can still occur in minor versions, if the alternative is to not
 
 * `AutocompleteSuggestion` has new `isAutocomplete` and `isUnknown` properties.
 * `AutocompleteSuggestionProvider` has new functions for ignoring and learning words.
+* `AutocompleteToolbarItem` is a new view that replicates a native autocomplete item.
+* `AutocompleteToolbarItemText` is a new view that replicates the text of a native autocomplete item.
 * `KeyboardLocale` now implementes `Identifiable`.
 * `KeyboardLocale` has new `flag`, `id` and `localeIdentifier` properties.
 * `KeyboardInputViewController` has a new `autocompleteSuggestionProvider` property.